recognition/requirement.txt
aiokafka==0.7.2
anyio==3.5.0
asgiref==3.5.1
click==8.1.3
fastapi==0.77.1
flatbuffers==2.0
grpcio==1.37.1
grpcio-tools==1.37.1
h11==0.13.0
idna==3.3
imageio==2.19.1
joblib==1.1.0
kafka-python==2.0.2
mmh3==3.0.0
networkx==2.8
numpy==1.22.3
onnx==1.11.0
onnxruntime==1.11.1
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
protobuf==3.20.1
pydantic==1.9.0
pymilvus==2.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyWavelets==1.3.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
starlette==0.19.1
threadpoolctl==3.1.0
tifffile==2022.5.4
typing_extensions==4.2.0
ujson==5.1.0
uvicorn==0.17.6
